Disentangle BVH and move into model with public BVH API
