--- newton/_src/sensors/warp_raytrace/bvh.py
@@ -1,227 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 The Newton Developers
-# SPDX-License-Identifier: Apache-2.0
-
-import warp as wp
-
-from ...core import MAXVAL
-from ...geometry import GeoType
-
-
-@wp.func
-def compute_shape_bounds(
-    transform: wp.transformf, scale: wp.vec3f, shape_min_bounds: wp.vec3f, shape_max_bounds: wp.vec3f
-) -> tuple[wp.vec3f, wp.vec3f]:
-    shape_min_bounds = wp.cw_mul(shape_min_bounds, scale)
-    shape_max_bounds = wp.cw_mul(shape_max_bounds, scale)
-
-    min_bound = wp.vec3f(MAXVAL)
-    max_bound = wp.vec3f(-MAXVAL)
-
-    corner_1 = wp.transform_point(transform, wp.vec3f(shape_min_bounds[0], shape_min_bounds[1], shape_min_bounds[2]))
-    min_bound = wp.min(min_bound, corner_1)
-    max_bound = wp.max(max_bound, corner_1)
-
-    corner_2 = wp.transform_point(transform, wp.vec3f(shape_max_bounds[0], shape_min_bounds[1], shape_min_bounds[2]))
-    min_bound = wp.min(min_bound, corner_2)
-    max_bound = wp.max(max_bound, corner_2)
-
-    corner_3 = wp.transform_point(transform, wp.vec3f(shape_max_bounds[0], shape_max_bounds[1], shape_min_bounds[2]))
-    min_bound = wp.min(min_bound, corner_3)
-    max_bound = wp.max(max_bound, corner_3)
-
-    corner_4 = wp.transform_point(transform, wp.vec3f(shape_min_bounds[0], shape_max_bounds[1], shape_min_bounds[2]))
-    min_bound = wp.min(min_bound, corner_4)
-    max_bound = wp.max(max_bound, corner_4)
-
-    corner_5 = wp.transform_point(transform, wp.vec3f(shape_min_bounds[0], shape_min_bounds[1], shape_max_bounds[2]))
-    min_bound = wp.min(min_bound, corner_5)
-    max_bound = wp.max(max_bound, corner_5)
-
-    corner_6 = wp.transform_point(transform, wp.vec3f(shape_max_bounds[0], shape_min_bounds[1], shape_max_bounds[2]))
-    min_bound = wp.min(min_bound, corner_6)
-    max_bound = wp.max(max_bound, corner_6)
-
-    corner_7 = wp.transform_point(transform, wp.vec3f(shape_min_bounds[0], shape_max_bounds[1], shape_max_bounds[2]))
-    min_bound = wp.min(min_bound, corner_7)
-    max_bound = wp.max(max_bound, corner_7)
-
-    corner_8 = wp.transform_point(transform, wp.vec3f(shape_max_bounds[0], shape_max_bounds[1], shape_max_bounds[2]))
-    min_bound = wp.min(min_bound, corner_8)
-    max_bound = wp.max(max_bound, corner_8)
-
-    return min_bound, max_bound
-
-
-@wp.func
-def compute_box_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    min_bound = wp.vec3f(MAXVAL)
-    max_bound = wp.vec3f(-MAXVAL)
-
-    for x in range(2):
-        for y in range(2):
-            for z in range(2):
-                local_corner = wp.vec3f(
-                    size[0] * (2.0 * wp.float32(x) - 1.0),
-                    size[1] * (2.0 * wp.float32(y) - 1.0),
-                    size[2] * (2.0 * wp.float32(z) - 1.0),
-                )
-                world_corner = wp.transform_point(transform, local_corner)
-                min_bound = wp.min(min_bound, world_corner)
-                max_bound = wp.max(max_bound, world_corner)
-
-    return min_bound, max_bound
-
-
-@wp.func
-def compute_sphere_bounds(pos: wp.vec3f, radius: wp.float32) -> tuple[wp.vec3f, wp.vec3f]:
-    return pos - wp.vec3f(radius), pos + wp.vec3f(radius)
-
-
-@wp.func
-def compute_capsule_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    radius = size[0]
-    half_length = size[1]
-    extent = wp.vec3f(radius, radius, half_length + radius)
-    return compute_box_bounds(transform, extent)
-
-
-@wp.func
-def compute_cylinder_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    radius = size[0]
-    half_length = size[1]
-    extent = wp.vec3f(radius, radius, half_length)
-    return compute_box_bounds(transform, extent)
-
-
-@wp.func
-def compute_cone_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    extent = wp.vec3f(size[0], size[0], size[1])
-    return compute_box_bounds(transform, extent)
-
-
-@wp.func
-def compute_plane_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    # If plane size is non-positive, treat as infinite plane and use a large default extent
-    size_scale = wp.max(size[0], size[1]) * 2.0
-    if size[0] <= 0.0 or size[1] <= 0.0:
-        size_scale = 1000.0
-
-    min_bound = wp.vec3f(MAXVAL)
-    max_bound = wp.vec3f(-MAXVAL)
-
-    for x in range(2):
-        for y in range(2):
-            local_corner = wp.vec3f(
-                size_scale * (2.0 * wp.float32(x) - 1.0),
-                size_scale * (2.0 * wp.float32(y) - 1.0),
-                0.0,
-            )
-            world_corner = wp.transform_point(transform, local_corner)
-            min_bound = wp.min(min_bound, world_corner)
-            max_bound = wp.max(max_bound, world_corner)
-
-    extent = wp.vec3f(0.1)
-    return min_bound - extent, max_bound + extent
-
-
-@wp.func
-def compute_ellipsoid_bounds(transform: wp.transformf, size: wp.vec3f) -> tuple[wp.vec3f, wp.vec3f]:
-    extent = wp.vec3f(wp.abs(size[0]), wp.abs(size[1]), wp.abs(size[2]))
-    return compute_box_bounds(transform, extent)
-
-
-@wp.kernel(enable_backward=False)
-def compute_shape_bvh_bounds(
-    shape_count_enabled: wp.int32,
-    world_count: wp.int32,
-    shape_world_index: wp.array[wp.int32],
-    shape_enabled: wp.array[wp.uint32],
-    shape_types: wp.array[wp.int32],
-    shape_sizes: wp.array[wp.vec3f],
-    shape_transforms: wp.array[wp.transformf],
-    shape_bounds: wp.array2d[wp.vec3f],
-    out_bvh_lowers: wp.array[wp.vec3f],
-    out_bvh_uppers: wp.array[wp.vec3f],
-    out_bvh_groups: wp.array[wp.int32],
-):
-    tid = wp.tid()
-    bvh_index_local = tid % shape_count_enabled
-    if bvh_index_local >= shape_count_enabled:
-        return
-
-    shape_index = shape_enabled[bvh_index_local]
-
-    world_index = shape_world_index[shape_index]
-    if world_index < 0:
-        world_index = world_count + world_index
-
-    if world_index >= world_count:
-        return
-
-    transform = shape_transforms[shape_index]
-    size = shape_sizes[shape_index]
-    geom_type = shape_types[shape_index]
-
-    lower = wp.vec3f()
-    upper = wp.vec3f()
-
-    if geom_type == GeoType.SPHERE:
-        lower, upper = compute_sphere_bounds(wp.transform_get_translation(transform), size[0])
-    elif geom_type == GeoType.CAPSULE:
-        lower, upper = compute_capsule_bounds(transform, size)
-    elif geom_type == GeoType.CYLINDER:
-        lower, upper = compute_cylinder_bounds(transform, size)
-    elif geom_type == GeoType.CONE:
-        lower, upper = compute_cone_bounds(transform, size)
-    elif geom_type == GeoType.PLANE:
-        lower, upper = compute_plane_bounds(transform, size)
-    elif geom_type == GeoType.ELLIPSOID:
-        lower, upper = compute_ellipsoid_bounds(transform, size)
-    elif geom_type == GeoType.BOX:
-        lower, upper = compute_box_bounds(transform, size)
-    elif geom_type == GeoType.MESH or geom_type == GeoType.GAUSSIAN:
-        min_bounds = shape_bounds[shape_index, 0]
-        max_bounds = shape_bounds[shape_index, 1]
-        lower, upper = compute_shape_bounds(transform, size, min_bounds, max_bounds)
-
-    out_bvh_lowers[bvh_index_local] = lower
-    out_bvh_uppers[bvh_index_local] = upper
-    out_bvh_groups[bvh_index_local] = world_index
-
-
-@wp.kernel(enable_backward=False)
-def compute_particle_bvh_bounds(
-    num_particles: wp.int32,
-    world_count: wp.int32,
-    particle_world_index: wp.array[wp.int32],
-    particle_position: wp.array[wp.vec3f],
-    particle_radius: wp.array[wp.float32],
-    out_bvh_lowers: wp.array[wp.vec3f],
-    out_bvh_uppers: wp.array[wp.vec3f],
-    out_bvh_groups: wp.array[wp.int32],
-):
-    tid = wp.tid()
-    bvh_index_local = tid % num_particles
-    if bvh_index_local >= num_particles:
-        return
-
-    particle_index = bvh_index_local
-
-    world_index = particle_world_index[particle_index]
-    if world_index < 0:
-        world_index = world_count + world_index
-
-    if world_index >= world_count:
-        return
-
-    lower, upper = compute_sphere_bounds(particle_position[particle_index], particle_radius[particle_index])
-
-    out_bvh_lowers[bvh_index_local] = lower
-    out_bvh_uppers[bvh_index_local] = upper
-    out_bvh_groups[bvh_index_local] = world_index
-
-
-@wp.kernel(enable_backward=False)
-def compute_bvh_group_roots(bvh_id: wp.uint64, out_bvh_group_roots: wp.array[wp.int32]):
-    tid = wp.tid()
-    out_bvh_group_roots[tid] = wp.bvh_get_group_root(bvh_id, tid)
