+++ newton/tests/test_sensor_tiled_camera.py
@@ -109,8 +109,11 @@ class TestSensorTiledCamera(unittest.TestCase):
         color_image = tiled_camera_sensor.utils.create_color_image_output(width, height, camera_count)
         depth_image = tiled_camera_sensor.utils.create_depth_image_output(width, height, camera_count)
 
+        state = model.state()
+        newton.geometry.build_shape_bvh(model, state)
+        newton.geometry.build_particle_bvh(model, state)
         tiled_camera_sensor.update(
-            model.state(), camera_transforms, camera_rays, color_image=color_image, depth_image=depth_image
+            state, camera_transforms, camera_rays, color_image=color_image, depth_image=depth_image
         )
 
         golden_color_data = np.load(
@@ -138,33 +141,33 @@ class TestSensorTiledCamera(unittest.TestCase):
         tiled_camera_sensor = SensorTiledCamera(model=model)
         camera_rays = tiled_camera_sensor.utils.compute_pinhole_camera_rays(width, height, math.radians(45.0))
 
+        state = model.state()
+        newton.geometry.build_shape_bvh(model, state)
+        newton.geometry.build_particle_bvh(model, state)
+
         color_image = tiled_camera_sensor.utils.create_color_image_output(width, height, camera_count)
         depth_image = tiled_camera_sensor.utils.create_depth_image_output(width, height, camera_count)
         tiled_camera_sensor.update(
-            model.state(), camera_transforms, camera_rays, color_image=color_image, depth_image=depth_image
+            state, camera_transforms, camera_rays, color_image=color_image, depth_image=depth_image
         )
         self.assertTrue(np.any(color_image.numpy() != 0), "Color image should contain rendered data")
         self.assertTrue(np.any(depth_image.numpy() != 0), "Depth image should contain rendered data")
 
         color_image = tiled_camera_sensor.utils.create_color_image_output(width, height, camera_count)
         depth_image = tiled_camera_sensor.utils.create_depth_image_output(width, height, camera_count)
-        tiled_camera_sensor.update(
-            model.state(), camera_transforms, camera_rays, color_image=color_image, depth_image=None
-        )
+        tiled_camera_sensor.update(state, camera_transforms, camera_rays, color_image=color_image, depth_image=None)
         self.assertTrue(np.any(color_image.numpy() != 0), "Color image should contain rendered data")
         self.assertFalse(np.any(depth_image.numpy() != 0), "Depth image should NOT contain rendered data")
 
         color_image = tiled_camera_sensor.utils.create_color_image_output(width, height, camera_count)
         depth_image = tiled_camera_sensor.utils.create_depth_image_output(width, height, camera_count)
-        tiled_camera_sensor.update(
-            model.state(), camera_transforms, camera_rays, color_image=None, depth_image=depth_image
-        )
+        tiled_camera_sensor.update(state, camera_transforms, camera_rays, color_image=None, depth_image=depth_image)
         self.assertFalse(np.any(color_image.numpy() != 0), "Color image should NOT contain rendered data")
         self.assertTrue(np.any(depth_image.numpy() != 0), "Depth image should contain rendered data")
 
         color_image = tiled_camera_sensor.utils.create_color_image_output(width, height, camera_count)
         depth_image = tiled_camera_sensor.utils.create_depth_image_output(width, height, camera_count)
-        tiled_camera_sensor.update(model.state(), camera_transforms, camera_rays, color_image=None, depth_image=None)
+        tiled_camera_sensor.update(state, camera_transforms, camera_rays, color_image=None, depth_image=None)
         self.assertFalse(np.any(color_image.numpy() != 0), "Color image should NOT contain rendered data")
         self.assertFalse(np.any(depth_image.numpy() != 0), "Depth image should NOT contain rendered data")
